Completata la creazione dei file SIAN per trasmettere le anagrafiche dei clienti e fornitori
git-svn-id: https://svn.code.sf.net/p/gazie/code/trunk@3415 0ed6177e-08db-47c9-9b37-4a62b3198fc7

+++ modules/camp/admin_sian_files.php
@@ -32,13 +32,26 @@ $admin_aziend = checkAdmin();
 require("../../library/include/header.php");
 $script_transl = HeadMain();
 
-if (isset ($_POST['confirm'])){
+if (isset ($_POST['confirm'])){ // cancello il file dalla cartella di GAzie
 	$filetodelete="../../data/files/".$admin_aziend['codice']."/sian/".$_POST['confirm'];
-	unlink ($filetodelete);
-	unset ($_POST,$form);
+	if (substr($_POST['confirm'],-12) == "OPERREGI.txt"){
+		unlink ($filetodelete);
+		unset ($_POST,$form);
+	}
+	if (substr($_POST['confirm'],-12) == "ANAGFCTO.txt"){
+		$fileContent=@file_get_contents($filetodelete); // prendo il contenuto del file
+		$filerecord=str_split ($fileContent,363);
+		foreach ($filerecord as $record) { // prendo l'ID anagrafica corrispondente 
+			$ana = gaz_dbi_get_row($gTables['anagra'],"codfis",substr($record,20,16));
+			$id=$ana['id'];// rimetto a zero lo status_SIAN
+			gaz_dbi_put_query($gTables['clfoco'],"id_anagra=$id","status_SIAN","0");
+		}
+		unlink ($filetodelete);
+		unset ($_POST,$form);
+	}
 }
 
-if (isset ($_POST['confirmsian'])){
+if (isset ($_POST['confirmsian'])){ // creo il file di cancellazione da inviare al SIAN
 	$filetodelete="../../data/files/".$admin_aziend['codice']."/sian/".$_POST['confirmsian'];
 	$fileContent=@file_get_contents($filetodelete); // prendo il contenuto dell'ultimo file da modificare
 	$fileField=explode (";",$fileContent);
@@ -94,9 +107,9 @@ if ($handle = opendir('../../data/files/1/sian/')){
 		  <table class="col-md-12 table-bordered table-striped table-condensed cf">
 		<thead class="cf">
 	<tr>
-	<th class="col-md-8">Nome file</th>
-	<th class="col-md-3">Giorno di creazione</th>
-	<th class="col-md-1">tipo</th>
+	<th class="col-md-7">Nome file</th>
+	<th class="col-md-2">Giorno di creazione</th>
+	<th class="col-md-2">tipo</th>
 	<th class="col-md-1">Scarica</th>
 	</tr>
 	</thead>
@@ -106,7 +119,7 @@ if ($handle = opendir('../../data/files/1/sian/')){
 			?>
 			<tr>
 				<td class="bg-danger">Conferma la cancellazione da GAzie di
-				<input type="submit" name="confirm" value="<?php echo $form['delete']; ?>" class="btn btn-xs btn-default btn-elimina">
+				<input type="submit" name="confirm" title="clicca per confermare" value="<?php echo $form['delete']; ?>" class="btn btn-xs btn-default btn-elimina">
 				</td>
 				<td>
 				<input type="submit" name="null" value="Annulla">
@@ -117,8 +130,8 @@ if ($handle = opendir('../../data/files/1/sian/')){
 	if (strlen($form['deletesian'])>0){
 			?>
 			<tr>
-				<td class="bg-danger">Conferma di voler creare il file SIAN per cancellare questo:
-				<input type="submit" name="confirmsian" value="<?php echo $form['deletesian']; ?>" class="btn btn-xs btn-default btn-elimina">
+				<td class="bg-warning">Conferma di voler creare un file SIAN che annulli questo:
+				<input type="submit" name="confirmsian" title="clicca per confermare" value="<?php echo $form['deletesian']; ?>" class="btn btn-xs btn-default btn-elimina">
 				</td>
 				<td>
 				<input type="submit" name="null" value="Annulla">
@@ -131,7 +144,16 @@ if ($handle = opendir('../../data/files/1/sian/')){
 			$filetoread="../../data/files/".$admin_aziend['codice']."/sian/".$file;
 			$fileContent=@file_get_contents($filetoread); // prendo il contenuto del file
 			$fileField=explode (";",$fileContent);
-			$filetype=$fileField[48];
+			if (substr($filetoread,-12) == "OPERREGI.txt"){
+				if ($fileField[48]=="I"){
+					$filetype="Inser. mov.";
+				} else {
+					$filetype="Canc. mov.";
+				}
+			} 
+			if (substr($filetoread,-12) == "ANAGFCTO.txt"){
+				$filetype="Anagr.";
+			}
 			$data=explode("_",$file);
 			$gio = substr($data[1],6,2);
 			$mes = substr($data[1],4,2);
@@ -142,32 +164,32 @@ if ($handle = opendir('../../data/files/1/sian/')){
 				<td data-title="Giorno"><?php echo $gio,"-",$mes,"-",$ann;?></td>
 				<td data-title="Tipo"><?php echo $filetype;?></td>
 				<td data-title="Scarica">
-				<a href="../camp/getfilesian.php?filename=<?php echo substr($file,0,-4);?>&ext=txt&company_id=1">
+				<a href="../camp/getfilesian.php?filename=<?php echo substr($file,0,-4);?>&folder=&ext=txt&company_id=1">
 				<i class="glyphicon glyphicon-file" title="Scarica il file appena generato">
 				</i></a></td>
 				<?php
 				if (!isset($first)){
 					?>
 					<td align="center">
-					<button type="submit" onclick = "this.form.submit();" name="del" value="del" class="btn btn-xs btn-default btn-elimina" >
-					<span class="glyphicon glyphicon-remove"></span>
+					<button type="submit" onclick = "this.form.submit();" title="Elimina da GAzie" name="del" value="del" class="btn btn-xs btn-default btn-elimina" >
+					<span class="glyphicon glyphicon-trash"></span>
 					</button>
 					</td>
 					<input type="hidden" name="first" value="<?php echo $file;?>">
-					
+					<?php if (substr($filetoread,-12) == "OPERREGI.txt"){ ?>
 					<td align="center">
-					<button type="submit" onclick = "this.form.submit();" name="del" value="delsian" class="btn btn-xs btn-default" >
+					<button type="submit" onclick = "this.form.submit();" title="crea file di cancellazione" name="del" value="delsian" class="btn btn-xs btn-default" >
 					<span class="glyphicon glyphicon-remove"></span>
 					</button>
 					</td>
 					<input type="hidden" name="first" value="<?php echo $file;?>">
+					<?php }?>
 				<?php
 					$first=1;
 				} 
 				?>
 			</tr>
-			<?php
-		
+			<?php	
 		}
 	}
 	?>
